Customise data entry, dataEntry component, fix test errors, change help message
