+++ src/pages/actual-values/ActualValues.tsx
@@ -3,32 +3,34 @@ import i18n from "../../locales";
 import PageHeader from "../../components/page-header/PageHeader";
 import { useHistory } from "react-router";
 import { History } from "history";
-import TargetValues from "../target-values/TargetValues";
+import DataEntry from "../../components/data-entry/DataEntry";
 
 function goTo(history: History, url: string) {
     history.push(url);
 }
 
-const DataEntry: React.FC = () => {
+const ActualValues: React.FC = () => {
     const history = useHistory();
     const goBack = () => goTo(history, "/projects");
-    const help = i18n.t(
-        `Select a) organizational unit where project was performed, b) data set, c) date, d) current/target
-
-        Then enter data for the fields shown in the screen.`
-    );
+    const help = i18n.t(`This is an example of help message.`);
     const subtitle = i18n.t(
         `Once cells turn into green, all information is saved and you can leave the Data Entry Section`
     );
+    const orgUnitId = "YuQRtpLP10I";
+    const datasetId = "BfMAe6Itzgt";
 
     return (
         <React.Fragment>
-            <PageHeader title={i18n.t("Data Entry")} help={help} onBackClick={goBack} />
+            <PageHeader
+                title={i18n.t("Set Actual Values for Project")}
+                help={help}
+                onBackClick={goBack}
+            />
             <div style={stylesSubtitle}>{subtitle}</div>
-            <TargetValues />
+            <DataEntry orgUnitId={orgUnitId} datasetId={datasetId} />
         </React.Fragment>
     );
 };
 const stylesSubtitle = { marginBottom: 10, marginLeft: 15 };
 
-export default DataEntry;
+export default ActualValues;
