+++ src/pages/root/Root.tsx
@@ -3,8 +3,8 @@ import { Route, Switch, HashRouter } from "react-router-dom";
 import LandingPage from "../landing-page/LandingPage";
 import ProjectsList from "../projects-list/ProjectsList";
 import Report from "../report/Report";
-import DataEntry from "../data-entry/DataEntry";
-import TargetValuesPage from "../target-values/TargetValuesPage";
+import ActualValues from "../actual-values/ActualValues";
+import TargetValues from "../target-values/TargetValues";
 import Dashboard from "../dashboard/Dashboard";
 import { generateUrl } from "../../router";
 import ProjectWizard from "../project-wizard/ProjectWizard";
@@ -16,8 +16,8 @@ const Root = () => {
                 <Route path={generateUrl("projects.new")} render={() => <ProjectWizard />} />
                 <Route path={generateUrl("projects")} render={() => <ProjectsList />} />
                 <Route path={generateUrl("report")} render={() => <Report />} />
-                <Route path={generateUrl("dataEntry")} render={() => <DataEntry />} />
-                <Route path={generateUrl("targetValues")} render={() => <TargetValuesPage />} />
+                <Route path={generateUrl("actualValues")} render={() => <ActualValues />} />
+                <Route path={generateUrl("targetValues")} render={() => <TargetValues />} />
                 <Route path={generateUrl("dashboard")} render={() => <Dashboard />} />
                 <Route render={() => <LandingPage />} />
             </Switch>
