+++ src/router.ts
@@ -3,7 +3,7 @@ const routes = {
     "projects.new": () => `/projects/new`,
     "projects.edit": ({ id }: { id: string }) => `/projects/${id}`,
     report: () => "/report",
-    dataEntry: () => "/data-entry",
+    actualValues: () => "/actual-values",
     // "dataEntry.edit": ({ id }: { id: string }) => `/dataEntry/${id}`,
     targetValues: () => "/target-values",
     // "targetValues.edit": ({ id }: { id: string }) => `/targetValues/${id}`,
